Set up erb_lint (#6)
This sets up ERB linting for views. The configuration
takes into account that we are using standardrb for
Ruby linting: 

https://thoughtbot.com/blog/a-standard-way-to-lint-your-views

The vendors directory is excluded to account for the
way the gems are installed on CI.

This also sets up a GitHub action to run the linter
on pull requests.