Add route constraints depending on site subdomain (#11)
This will allow the sign in routes and any other admin-only pages
only to be accessed when on an admin subdomain. The routes
will return a 404 if accessed on a non-admin subdomain. The
dev setup can be accessed at http://admin.example.localhost:3000
as long as /etc/hosts is configured correctly.

These changes also add different root paths depending on
whether the user is signed in or not, or whether the user is
on a non-admin subdomain.