Add active storage attachment factory that does not create file_import record through callback
When creating a file_record object through the
factory, it was creating an active_storage_attachment
record, which in turn was creating another file_import
record through the after_commit callback. This adds
an active_storage_attachment factory that suppresses
the functionality of the callback for when we are
creating file_import records through the file_import
factory.

An addition to the system spec is to make sure
to test content that only appears on the page the
user is being redirected to, before checking for the
status of the file_import record. It's possible that
the spec could sometimes use the File Imports text in
the nav to consider that the spec was successful, and
then we would end up with a race condition where the
database update hadn't always completed when checking
the new file status.