Allow editing of already converted data_import files (#79)
This change allows a data_import file to be updated. When the
data_import file is replaced, it will destroy all the existing
associations on the occupation standard record to ensure that
we don't end up with work processes, related instructions, or
wage steps from the previous upload.

This also cleans up the import services to remove the returning
of the objects since they aren't being used.