Notify admin when guest uploads new standards import
This removes the code that would notify an admin
on every new standards import (created from
scrapers, guest page, or internal admin upload).

The converters are currently actively working
on the site and do not need to be notified when
new files are added through the scrapers. In the
future we can add a separate mailer that gets
triggered when the scrapers add files.