Setup elasticsearch models (#225)
Add a Brewfile and calling it from the setup script. This means elasticsearch will automatically be installed via brew once we're setting up the app. 

Add a Searchable module, which is included on the Occupation Standard model. This allows us to search on occupation standards. 

And add the elasticsearch service to our github workflow.