Implement the "More Like this" query for occupation standards (#255)
Refactor Previous work: We don't need the searchable and occupation standard search modules anymore. We'll only need the SimilarOccupationStandards class;

Include the elasticsearch-rails and elasticsearch-model gems;

Implement a "More Like This" ES query to return similar occupation standards.

"The More Like This Query finds documents that are "like" a given set of documents. In order to do so, MLT selects a set of representative terms of these input documents, forms a query using these terms, executes the query and returns the results. The user controls the input documents, how the terms should be selected and how the query is formed."