Move some words to acronyms
"RAPIDS" was displaying as "R API D"
on the Administrate index column
header due to "API" being an acronym.
This adds RAPIDS to the acronyms list,
along with a few other additions. Notably,
ONET was **not** added as there is an
Onet model, and this was going to involve
a lot of code changes to rename everything
to ONET.