Handle registration agency display for National agency (#292)
We recently added a registration agency with no
associated state to represent a national
registration agency (4a71ab). However, this was
causing the edit occupation standard page to break,
due to the lack of a state association. This will
now display "National (OA)" for the national agency
in the select options.