Improve the daily uploads report mailer (#302)
* Improve the daily uploads report mailer

In addition to the min/max for OJT hours and RSI hours, also display the occupation_standard.work_processes_hours and occupation_standard.related_instructions_hours counts which are used on the website.
Only show a line if we have data for it. ie, hide Competencies count if 0, hide any “RSI hours min” if nil, etc.
Reorganize the links to shorten up the email.