Remove more_like_this in Elasticsearch query for similar_to method
The built-in more_like_this feature in Elasticsearch
seems better suited to finding documents related to
some simple text, or a group of input documents.

Per the Alternative section in the Elasticsearch
documentation, a more complex boolean query can
be used to achieve the same functionality of
more_like_this, with more control:

"To take more control over the construction of a query
for similar documents it is worth considering writing
custom client code to assemble selected terms from an
example document into a Boolean query with the desired
settings."
Ref: https://www.elastic.co/guide/en/elasticsearch/reference/7.17/query-dsl-mlt-query.html#_alternative

We already have such a boolean query established, so
this removes the unnecessary additional call to the
more_like_this function.