Hide accordion for production
We need to setup the condensing of occupation standards so that similar programs will be grouped. That way the accordion showing similar programs will make sense. Currently it' redundant because the same results inside the accordion are also listed in the page with their own accordions. I hid the accordion unless we are in tests so that the tests could remain. Once we condense the results we can add back the accordion.