Remove national_standards controller (#382)
The national standards controller
is returning a subset of the occupation
standards, but this configuration
is no longer necessary. Now that we
have filtering properly set up, the
National link should return results
as if the national standards checkboxes
were checked in the filters.

This PR is in preparation for handling
all searching through Elasticsearch.