Move search to Elasticsearch (#383)
This sets up the option to run all of the searching
and filtering through Elasticsearch. 

The analyzer for the occupation standard title and
industry name was modified to use ngrams so that
we could more easily do partial matching. The new
analyzer uses the English stop words and 3-character
n-grams.

Several system specs have been moved over to view
specs instead.