Fix Elasticsearch searching  (#406)
The Elasticsearch query was trying to fit both the needs
for autocomplete and regular stemming into the same
search query, which was resulting in an overmatch of records.
This simplifies the title, related job titles, and industry name
queries to use the English analyzer, which handles stop words,
lowercasing, and stemming. In addition, we add a separate
index on the title field to use the autocomplete analyzer so
that the typeahead feature works correctly. This also adds a
match phrase query on the title with a higher boost value than
the regular search on the title field.