Refine Elasticsearch results (#427)
When a user enters a search term, they are often returned
results unrelated to their intended search due to related words.
For example, searching "software developer" will also return
results for "Child Development Teacher" because of the word
"developer" in the search. To refine the elasticsearch results,
when a user enters a search term, we look up the ONET code
of the top hit and do a second search that only returns results
matching the ONET prefix. In that way, "Child Development Teacher"
would be excluded from a search for "software developer" since
they do not share the same ONET code prefix.