Set current version on existing onets (#443)
The current taxonomy in use for ONET SOC-codes is 2019.
This adds a version field to the onets table and populates
the current records with the "2019" version. This also makes
a slight modification to the ScrapeOnetCodes service to treat
the ONET code as the canonical source for the record, rather
than the title.