Move populating onets and occupations out of seeds file (#458)
The `bin/setup` task runs the seeds file, which in
turn was running the tasks to populate the onets
table and the occupations table. The onets population
task is particularly time-consuming now as it makes
a call to a 3rd-party API to populate the related
job titles fields. These tasks really only need
to be run once for dev environment setup, so they
are being moved to a one-time After Party task.