Delete index in begin/rescue before ES specs to prevent flakes
While we only run the before/after create/delete
index tasks on specs which have the :elasticsearch
flag, we still have Elasticsearch callbacks on
the Occupation and OccupationStandard models
which are automatically creating the index whenever
we create an Occupation or OccupationStandard instance.
These specs are therefore creating the ES index, but
never deleting it at the end of the spec. This change
adds a check at the beginning of the ES specs
to delete an index if it exists before beginning
the spec.