Eager load application in test environment (#484)
In the elasticsearch specs, the start up
for creating Elasticsearch models calls
ApplicationRecord.descendants. However, when
running a single spec, this does not
always load both Occupation and OccupationStandard
so the indexes may not be created before the start
of running the spec. This results in the index
getting created by default through the Elasticsearch
callbacks when the record is created, and for some
reason it does not apply the settings that are defined
in the model correctly.