Fixes and additions for Admin dashboard
* Add GenericRecord field type
There was no way to display the StandardsImport
associated with a SourceFile, since there
wasn't a built-in association between
the objects. This adds a new GenericRecord
administrate field that display a link
to the record.

* Fixed editing of StandardsImport
The StandardsImport uses custom
permitted_params but the latest
Administrate release had updated
that method to take a parameter
https://github.com/thoughtbot/administrate/pull/2356