Do not display expansion icon if nothing to display (#536)
Currently when viewing an occupation standard, the work
process and related instruction sections will show the red
plus icon to toggle content, even when there is no content
to display. The changes here refactor the word processes
and related instructions tabs to use partials to avoid redundant
code, and no longer displays empty details in the HTML
details/summary element when there is nothing to display.