Use aggregation of headline as correct total for occupation standards page numbering
The only Elasticsearch-related code in the Pagy
`new_from_elasticsearch_rails` method is code
used to compute the total number of pages. Since
that code uses the total number of hits, it is
incorrect for our usage. This change just
instantiates a new Pagy object passing in the
correct aggregated value for the total number of
records.