Method of retrieving duplicates depends on using ES for search or not
As we made the Expand Duplicates a permanent feature,
the query used to find those duplicates should now
depend on whether or not we are using Elasticsearch
for search or not. The similar_programs_elasticsearch
flag is for a separate query to return similar
programs to display on the occupation_standards#show
page.