Create ExportFileAttachments service for getting attachments out of docx files
To do so, we are passing a SourceFile object and a file blob for the
service. With those, we are able to transform the docx file into a zip
file, which is necessary for getting the attached files out. We then
unzip the file, getting a list of attached files and their respective
paths. Finally, we create a StandardImport based on the SourceFile
passed to the service, and attach each of the extracted files to it. As
a clean up stage, we unlink the tempfile created for the zip, and delete
the extracted files.