Refactor ExportFileAttachments service to become a job
We turned it into a job so that we could call it multiple times without
having the possibility of getting an exception and disrupting the whole
flow of the app. This will allow us to get visibility into which
exporting jobs failed, and have the advantage of performing then async.