Export and saves attachments out of docx file (#557)
* Create ExportFileAttachments service for getting attachments out of docx
files

To do so, we are passing a SourceFile object and a file blob for the
service. With those, we are able to transform the docx file into a zip
file, which is necessary for getting the attached files out. We then
unzip the file, getting a list of attached files and their respective
paths. Finally, we create a StandardImport based on the SourceFile
passed to the service, and attach each of the extracted files to it. As
a clean up stage, we unlink the tempfile created for the zip, and delete
the extracted files.

* Add updated Gemfile.lock

* Modify temp file path to include Rails root

* Extract attached files to tmp directory

* Cleanup and typos

Co-authored-by: Jeanine Soterwood <jeanine@thoughtbot.com>

* Refactor ExportFileAttachments service

Change the StandardImport url to admin_source_file_path, and get blob
from the source_file instead of initializing it with the class.

* Add treatment for files with no attachment

Doesn't try to open zip files that have no attachments to them.

* Add ExportFileAttachments service call to ApprenticeshipBulletinsJob

Also, refactored the service to not create a StandardImport when it
doesn't have files attached to it.

* Fix typo on ApprenticeshipBulletinsJob

* Fix specs

* Refactor ExportFileAttachments service to become a job

We turned it into a job so that we could call it multiple times without
having the possibility of getting an exception and disrupting the whole
flow of the app. This will allow us to get visibility into which
exporting jobs failed, and have the advantage of performing then async.

* Style changes

* Fix standardrb lint errors

---------

Co-authored-by: Jeanine Soterwood <jeanine@thoughtbot.com>