Add task to convert docx files to pdf's
This connects code from a few different areas to allow processing all
SourceFiles with docx attachments and converting them into pdf's. The
docx file is preserved as `source_file.active_storage_attachment` and
the new PDF is uploaded as `source_file.redacted_file`.

I didn't attempt to convert SourceFiles which
* already have a redacted file attachment, or
* have a docx file embedded attachments

Right now this is also just a script that you can run once. It's
idempotent so you should be able to run it over and over if need be, but
because we need to download the file to check if it has attachments it
will still try to reprocess those files each time you run the script. I
thought about adding a `converted_to_pdf_at` column or something but
opted not to for now. lmk if you feel differently though!

I think maybe a next step would be to run this in the background each
time a source file is created so that all happens automatically and we
can just delete the rake task eventually.