Create new SourceFile instead
Instead of attaching the newly created PDF to the original docx source
file as `redacted_file`, instead create a wholly new SourceFile and then
mark the old docx one as arcived.

This is a little awkward to do because source files are created
automatically via a callback whenever the StandardsImport is modified.
Since we're unable to pass arguments into that, instead we set a
`link_to_pdf_filename` field on the docx source file. In
`CreateSourceFilesJob` then, after the new pdf source file is created,
we look for any existing source files that match and link them up.

A potental future refactor that could simplify this a bit could be to
add a service object or something for creating source files rather than
making it a callback on StandardsImport. That way we could pass
arguments into it and just generally have a bit more control over when
it runs.