Add task to convert docx files to pdf's (#600)
* Add task to convert docx files to pdf's

This connects code from a few different areas to allow processing all
SourceFiles with docx attachments and converting them into pdf's. The
docx file is preserved as `source_file.active_storage_attachment` and
the new PDF is uploaded as `source_file.redacted_file`.

I didn't attempt to convert SourceFiles which
* already have a redacted file attachment, or
* have a docx file embedded attachments

Right now this is also just a script that you can run once. It's
idempotent so you should be able to run it over and over if need be, but
because we need to download the file to check if it has attachments it
will still try to reprocess those files each time you run the script. I
thought about adding a `converted_to_pdf_at` column or something but
opted not to for now. lmk if you feel differently though!

I think maybe a next step would be to run this in the background each
time a source file is created so that all happens automatically and we
can just delete the rake task eventually.

* Shorten trait names

Rename `with_pdf_attachment` to just `pdf` since all source files will
have an attachment. More than anything I did this because I wanted one
trait for a regular docx file and another for a docx file with
attachments, and there was already a `with_docx_attachment` trait, so
I would have needed to add something like
`with_docx_attachment_with_attachments` and that felt too confusing 😅

* Create new SourceFile instead

Instead of attaching the newly created PDF to the original docx source
file as `redacted_file`, instead create a wholly new SourceFile and then
mark the old docx one as arcived.

This is a little awkward to do because source files are created
automatically via a callback whenever the StandardsImport is modified.
Since we're unable to pass arguments into that, instead we set a
`link_to_pdf_filename` field on the docx source file. In
`CreateSourceFilesJob` then, after the new pdf source file is created,
we look for any existing source files that match and link them up.

A potental future refactor that could simplify this a bit could be to
add a service object or something for creating source files rather than
making it a callback on StandardsImport. That way we could pass
arguments into it and just generally have a bit more control over when
it runs.