Only create source files on standards_import update
Because standards import records have files attached
to them, the record is marked as updated after
the last file has been attached. This means that
the CreateSourceFilesJob is called twice in
quick succession, sometimes resulting in
duplicate source files. Since we only need
to run this job if files have been attached,
it is sufficient to run it on after_update.