Only create source files on standards_import update (#638)
Because standards import records have files attached
to them, the record is marked as updated after
the last file has been attached. This means that
the CreateSourceFilesJob is called twice in
quick succession, sometimes resulting in
duplicate source files. Since we only need
to run this job if files have been attached,
it is sufficient to run it on after_update.

Originally, the source file creation happened
as an after_create on the ActiveStorage::Attachment
record itself, but that changed with the Rails 7.1 upgrade.
Ref b150ee2