Create source file from active storage after commit hook (#656)
As part of the work for the docx -> pdf conversion, in particular the extraction
of embedded attachments from Bulletin Word docs, we are manually attaching
files to a standards_import record, meaning that the background job to create
source files will get called multiple times in quick succession every time a new
file is added to a a StandardsImport. This behavior is not desired and may cause
race conditions with extraneous source files getting created.

The new behavior is such that any time an ActiveStorage::Attachment record is
created that is linked to a StandardsImport, then we will generate an associated
SourceFile.