Account for .doc and .pdf files as bulletin file types (#680)
The existing code assumes that all Bulletins will be of .docx file
type, but this is not the case. Some Bulletins are of .doc format
and others are of .pdf format. These changes unify code to
consider a generic "word" filetype, and ensure that we do not
attempt to extract embedded objects from pdf files.