Bulletin DocToPdf conversion fixes (#688)
This adds a new field on the StandardsImport
model so we can explicitly flag a Bulletin document.
The Bulletins will either have no attachments (and no
useful information and can be archived) or will have
attachments, which will get extracted in the
Scraper::ExportFileAttachmentsJob. We add a check
on the SourceFile model to not bother sending any
bulletins through to the DocToPdfConverterJob since
it is not necessary and is affecting the file integrity.