Do not include nil source_files for a StandardsImport record (#702)
When new files are manually added to a StandardsImport record,
the associated source file records are not immediately available
since they are created in a background job. However, the source_files
method is returning an array that includes these nil source files, causing
issues with displaying the list of source files in the Administrate view.
This change modifies the source_files method to not return any nil values.