Imports::Doc#process
Process a Doc import file by converting it to PDF.

We want to track whether the processing has finished, so we add new
columns `processed_at` and `processing_errors`. Only one is present at a
time.

We can also start to make use of the `status` column, to flag import
files that need attention.

---

Much of the work for this is already done in `DocToPdfConverter`.
However, that service hardcodes the idea of a `SourceFile`. Instead, we
want a service that is divorced from the database entirely: it should
take a string representing some sort of unique ID, and the file to
process. On success we need the filename of the PDF output.

So to ease the transition, make a `ConvertDocToPdf` service and use it
from `DocToPdfConverter`. Many of the tests are identical, just simpler.

While here, remove the unused `DocToPdfConverter.convert_all` method,
which attempts to enqueue a job but that job misspelled `#perform`.

<https://app.asana.com/0/1203289004376659/1207002310256046>