DocxListing#process
To process a DocxListing (a Docx file with attachments), first
split the attachments out. For each attachment, create an
`Imports::Uncategorized` child, and then process all children.

Along the way we realized that we don't handle the case where the
Uncategorized file is of an unknown type. For that, we raise.

The DocxListing#process method does not handle the error; any invalid
attachment stops everything. If we ever start processing files with
unexpected types we can reconsider that (though honestly we'll probably
want to process those files instead).

<https://app.asana.com/0/1203289004376659/1207002310256045>