Add DocxListingSplitter, extract OLE storage
Extract a `DocxListingSplitter` from `ExportFileAttachmentsJob`. This
service is only in charge of turning an ActiveStorage file into its
attachments, and knows nothing of the database. We need this for a
future refactoring.

While working on this, we learned that the `.bin` files attached to Docx
docs are _not_ PDFs. They are OLE storage files
(`application/x-ole-storage`) which _contain_ PDFs. Since anything that
works with PDFs expects a PDF, passing the OLE storage file around has
variable results:

- macOS's Preview program can open OLE storage files containing PDFs,
  and treats them as if they are PDFs.
- MATE's Atril program does not know what to do with them.
- soffice is unable to convert them to PDF.
- Marcel, the MIME detector in Rails, reports that they are
  `application/x-ole-storage`.
- most crucially, `SourceFile.pdf_attachment` does not recognize them as
  PDFs (since it relies on Marcel), meaning these PDF attachments are
  effectively dropped.

The solution is to assume all `.bin` files are OLE storage files
wrapping PDFs. We can use the ruby-ole library to extract the `CONTENTS`
key and write that to a `.pdf` file, then the rest of the system can
ignore the `.bin`.

We assume all `.bin` files are PDFs because so far this has been the
case. I tried to find an OLE storage mechanism for knowing the MIME type
of the attachments but didn't find it.