Do not perpetuate listing param beyond uncategorized imports
The listing parameter is used to determine the
Import type for files that have embedded documents
within them. Currently this parameter is continued
to be passed when processing children of DocxListing
records, which means we are attempting to unzip
regular docx files, resulting in empty zip files and
errors when trying to cleanup non-existent files.
Once the listing parameter has been used to determine
whether an uncategorized Import is a DocxListing or
not, then we can set the parameter to false since
no children of a DocxListing will ever be a DocxListing
record.