Add rake tasks to convert existing files to Import records (#790)
* Add temporary source_file reference to imports

While we transition from the old system to the new
system, it will be helpful to track which source_file
an Imports::Uncategorized record initially came from.
This field will be removed once we completely cut
over to the new tree system.

* Add rake task to create imports for source files

This takes all unarchived SourceFile records that do
not have a corresponding import record and creates
a new Imports::Uncategorized record that links to a
SourceFiles attachment. 

* Add rake task to process unfurled imports

* Add rake task to set import field on data_import record

This rake task takes the existing source_file linked to
the data import and finds the pdf import leaf and 
associates it with the data import.