Do not create imports for standards_imports for now (#791)
We want to let the new Import process handle the tree creation
for bulletin StandardsImports from scratch, since we were not
handling the OLE file types correctly when we were initially
extracting the attachments. We will skip any source files that
were derived from a bulletin and write a new task to handle them
separately.