Handle case when uncategorized child is docx_listing when finding pdf_leaf (#796)
When an Imports::Uncategorized child is of type Imports::DocxListing,
it doesn't have a single pdf leaf. In these cases the Imports::DocxListing
record will have multiple children of type Imports::Uncategorized which
will have the pdf leaf. If an Imports::Uncategorized record is the parent
of a Imports::DocxListing record, then just return nil for the pdf_leaf.