Remove extracted attachments from bulletins with no processing
There are some bulletins in which the extracted
attachments have been added to the standards_import,
but none of these source files have been processed.
In these cases, remove all of the extracted attachments
from the standards_import. This will leave only the
original file that was downloaded from the Bulletins
scraper job. The standards_import record is then
marked as bulletin: false so we can pick it up
to run it through the rake task that will create
imports for its source file record.