Do not create Uncategorized imports in files= method
It turns out the ActiveStorage `attach` method calls
`files=` under the hood. This means that we are
creating duplicate Imports::Uncategorized records
all over the place. However, we fortunately put
the incorrect status of `pending` on the duplicate
Uncategorized records, so they never got processed
and will be easy to clean up.

This changes the overriding of the `files=` method
for StandardsImport and just builds the Uncategorized
records in the controller itself. Since we ultimately
will NOT be attaching any files to the StandardsImport
records, it does not make sense to continue
to customize the `files=` method.