Allow x-ole-storage for .doc files to proceed as Doc files
There are several old .doc files that have
application/x-ole-storage content type. We
can see the WordDocument in the ole container
but when writing that to a file it becomes
corrupted.

However, when testing on both Mac and Linux
machines, we are able to open these files
directly as `.doc` files and also able to
convert them all the way through to PDF files.