Add invitation sent timestamp to user admin dashboard to prevent duplicate invitations (#967)
This change introduces a new field, `invitation_sent_at`, to the `administrator` user's dashboard,displaying the timestamp of when the invitation email was sent. By including this information on the user's show page in the admin panel, administrators will have a clearer view of whether an invitation has already been sent, reducing the risk of accidentally sending duplicate invitations.