Remove Occupation Standards create button on admin (#1001)
An Occupation Standard gets created when we import standards. A PDF file
is needed. The admin panel showed a "New Occupation Standard" button though
which fails when trying to open it as there's no way to upload a PDF.
This PR basically just overwrites the index_header partial used by
administrate to remove the new record creation button.
I did not add a system spec on purpose as we usually don't test what we
don't expect. There are tests for the Standard Import to work correctly.