Add dynamic work processes field to occupation standard (#1002)
Occupation Standards have many work processes. When we edit an occupation standard, we want to dynamically add and remove work processes without changing dashboards. During a first iteration we thought of using the rondo_form gem but since we are using administrate, we would need to go against rails conventions to make it work. The administrate-field-nested_has_many gem provides exactly the workflow we are looking for so I would suggest to move forward with this one instead of rondo_form.

This PR only adds the functionality to add and remove a work process. We will update the UI in a follow up PR.

In order to add or remove a competency to/from a work process via
the occupation standard admin page, we introduced a nested has_many
field for this record.
We also updated the factories for work processes and competencies
and added the system test for occupation standards.

Co-authored-by: Clarissa Borges <cborges@thoughtbot.com>