Update UI for nested has_many fields (#1010)
An occupation standard can have multiple work processes. While
we limit the display of these work processes on the index page, we
showed all of them, including their competencies, when we edited
an occupation standard. It was really difficult to navigate through
these fields. In order to make it easier, we used details and summaries
to show/hide the work processes and competencies.