Fix test for claim button
The system test for the claim button gave us a false
positive – we clicked the Claim button but we actually
did a "wrong" request: instead of submitting the form,
we linked to the show page of the import. This is related
to the upgrade of administrate where turbo got added.
I've added `js: true` to the test and ensured it would
fail without the fix.