Fix Claim button (#1022)
Fix Claim button

With the last upgrade of administrate, turbo got
introduced. The form for the claim button on imports
is nested in an anchor tag. With turbo enabled on links
by default, the link itself takes precedence over the
form. This led to the fact that we always got redirected
to the show page of an import but we were not able to
add an assignee. This commit fixes it.

The system test for the claim button gave us a false
positive – we clicked the Claim button but we actually
did a "wrong" request: instead of submitting the form,
we linked to the show page of the import. This is related
to the upgrade of administrate where turbo got added.
I've added `js: true` to the test and ensured it would
fail without the fix.