Fetch related instructions (#1024)
* Update OpenAI prompt

We want to also include related instructions when we use
ChatGPT to generate text from a PDF input. While working on
this, we found some issues with the way we create imports. A
clear PDF wasn't identified to be a PDF but had the uncategorized
type. Manually changing the type fixed it.

There's another issue regarding TooManyRequests which we will
pair on tomorrow.

* Update prompt

---------

Co-authored-by: Fernando Perales <fer@thoughtbot.com>